+++ run-blocking-core/commonTest/kotlin/com/javiersc/run/blocking/RunBlockingTest.kt
@@ -1,12 +1,12 @@
-package com.javiersc.runBlocking
+package com.javiersc.run.blocking
 
 import kotlin.test.Test
 import kotlin.test.assertTrue
 import kotlinx.coroutines.yield
 
-internal class SuspendTestTest {
+internal class RunBlockingTest {
 
-    @Test fun testSuspendTest() = suspendTest { assertTrue { getBoolean() } }
+    @Test fun testRunBlocking() = runBlocking { assertTrue { getBoolean() } }
 }
 
 internal suspend fun getBoolean(): Boolean = yield().run { true }