+++ run-blocking-core/jvmMain/kotlin/com/javiersc/run/blocking/RunBlocking.kt
@@ -1,10 +1,10 @@
-package com.javiersc.runBlocking
+package com.javiersc.run.blocking
 
 import kotlin.coroutines.CoroutineContext
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.runBlocking
 
-/** Native `runBlocking` implementation */
+/** JVM `runBlocking` implementation */
 public actual fun <T> runBlocking(
     context: CoroutineContext,
     block: suspend CoroutineScope.() -> T