+++ CHANGELOG.md
@@ -6,6 +6,11 @@
 
 ### Changed
 
+- `suspend-test` returns `TestResult`
+- `suspend-test` block only accepts `TestScope`
+- `suspend-test` used `runTest` under the hood
+- package to `com.javiersc.run.blocking`
+
 ### Deprecated
 
 ### Removed
@@ -14,7 +19,7 @@
 
 ### Updated
 
-- `com.javiersc.gradle-plugins:all-plugins -> 0.1.0-rc.28`
+- `com.javiersc.gradle-plugins:all-plugins -> 0.1.0-rc.33`
 - `org.jetbrains.kotlinx:kotlinx-coroutines-core -> 1.6.0`
 - `gradle -> 7.3.3`
 - `org.jetbrains.kotlin:kotlin-gradle-plugin -> 1.6.10`